third_party/blink/web_tests/virtual/http-cache-default-backend/external/wpt/fetch/http-cache/heuristic.any.sharedworker-expected.txt

This is a testharness.js-based test.
[FAIL] HTTP cache reuses an unknown response with Last-Modified based upon heuristic freshness when Cache-Control: public is present
  assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
[FAIL] HTTP cache reuses a 204 No Content response with Last-Modified based upon heuristic freshness
  assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
[FAIL] HTTP cache reuses a 404 Not Found response with Last-Modified based upon heuristic freshness
  assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
[FAIL] HTTP cache reuses a 405 Method Not Allowed response with Last-Modified based upon heuristic freshness
  assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
[FAIL] HTTP cache reuses a 414 URI Too Long response with Last-Modified based upon heuristic freshness
  assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
[FAIL] HTTP cache reuses a 501 Not Implemented response with Last-Modified based upon heuristic freshness
  assert_less_than: Response 2 does not come from cache expected a number less than 2 but got 2
Harness: the test ran to completion.
